backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
ffmpeg-python
sentence-transformers
transformers
hdbscan
umap-learn
yake
python-dotenv
requests
